style-transfer/pyproject.toml
[tool.poetry]
name = "style-transfer"
version = "0.1.0"
description = "Neural style transfer with PyTorch"
authors = ["You"]
packages = [{ include = "style_transfer", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
torch = "^2.0.0"
torchvision = "^0.15.0"
pillow = "^10.0.0"
matplotlib = "^3.8.0"
tqdm = "^4.66.0"
streamlit = "^1.32.0"
numpy = "^1.26.0"
mlflow = "^2.22.0"
mlflow-skinny = "^2.22.0"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
python-multipart = "^0.0.20"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
